# codeaware-py/openapi/current-release.json
{
  "components": {
    "schemas": {
      "AiOperationRecordRead": {
        "properties": {
          "ai_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Model"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "prompt_template_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Template Id"
          },
          "result": {
            "title": "Result",
            "type": "string"
          },
          "source_code": {
            "title": "Source Code",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "project_name",
          "file_path",
          "source_code",
          "result",
          "created_at"
        ],
        "title": "AiOperationRecordRead",
        "type": "object"
      },
...
      "Body_upload_file_api_knowledge_upload_file_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "PDF/DOCX/HTML/Markdown/TXT，原始文件最大 5 MiB",
            "title": "File",
            "type": "string"
          },
          "project_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_knowledge_upload_file_post",
        "type": "object"
      },
...
      "CodeReviewRequest": {
        "properties": {
          "file_path": {
            "maxLength": 500,
            "minLength": 1,
            "title": "File Path",
            "type": "string"
          },
          "project_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Project Name",
            "type": "string"
          },
          "prompt_template_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Template Id"
          },
          "source_code": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Source Code",
            "type": "string"
          }
        },
        "required": [
          "project_name",
          "file_path",
          "source_code"
        ],
        "title": "CodeReviewRequest",
        "type": "object"
      },
      "CodeReviewVO": {
        "properties": {
          "ai_model": {
            "default": "deepseek-v4-flash",
            "title": "Ai Model",
            "type": "string"
          },
          "critical_count": {
            "default": 0,
            "title": "Critical Count",
            "type": "integer"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "info_count": {
            "default": 0,
            "title": "Info Count",
            "type": "integer"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ReviewIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "issues_count": {
            "default": 0,
            "title": "Issues Count",
            "type": "integer"
          },
          "project_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          },
          "score": {
            "default": 0,
            "title": "Score",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "warning_count": {
            "default": 0,
            "title": "Warning Count",
            "type": "integer"
          }
        },
        "title": "CodeReviewVO",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KnowledgeDocumentVO": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "title": "KnowledgeDocumentVO",
        "type": "object"
      },
      "KnowledgeSearchHit": {
        "properties": {
          "chunk_content": {
            "title": "Chunk Content",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "match_type": {
            "enum": [
              "vector",
              "keyword",
              "both"
            ],
            "title": "Match Type",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "score",
          "match_type",
          "document_id",
          "chunk_content"
        ],
        "title": "KnowledgeSearchHit",
        "type": "object"
      },
      "KnowledgeSearchRequest": {
        "properties": {
          "query": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "KnowledgeSearchRequest",
        "type": "object"
      },
      "KnowledgeUploadRequest": {
        "properties": {
          "content": {
            "maxLength": 200000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "project_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          },
          "source_type": {
            "default": "MANUAL",
            "enum": [
              "MANUAL",
              "DOC"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "KnowledgeUploadRequest",
        "type": "object"
      },
...
      "PageResult_AiOperationRecordRead_": {
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/AiOperationRecordRead"
            },
            "title": "Records",
            "type": "array"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "size",
          "records"
        ],
        "title": "PageResult[AiOperationRecordRead]",
        "type": "object"
      },
...
      "Result_AiOperationRecordRead_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiOperationRecordRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[AiOperationRecordRead]",
        "type": "object"
      },
...
      "Result_CodeReviewVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CodeReviewVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[CodeReviewVO]",
        "type": "object"
      },
      "Result_KnowledgeDocumentVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeDocumentVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[KnowledgeDocumentVO]",
        "type": "object"
      },
...
      "Result_NoneType_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[NoneType]",
        "type": "object"
      },
      "Result_PageResult_AiOperationRecordRead__": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageResult_AiOperationRecordRead_"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[PageResult[AiOperationRecordRead]]",
        "type": "object"
      },
...
      "Result_list_KnowledgeSearchHit__": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/KnowledgeSearchHit"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[list[KnowledgeSearchHit]]",
        "type": "object"
      },
...
      "ReviewIssue": {
        "properties": {
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "dimension": {
            "minLength": 1,
            "title": "Dimension",
            "type": "string"
          },
          "fix_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fix Code"
          },
          "line_range": {
            "minLength": 1,
            "title": "Line Range",
            "type": "string"
          },
          "severity": {
            "minLength": 1,
            "title": "Severity",
            "type": "string"
          },
          "suggestion": {
            "minLength": 1,
            "title": "Suggestion",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "dimension",
          "severity",
          "line_range",
          "title",
          "description",
          "suggestion"
        ],
        "title": "ReviewIssue",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CodeAware - AI 驱动的研发效能平台 (Python 重构)",
    "title": "codeaware",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/code-review/records": {
      "get": {
        "operationId": "list_records_api_code_review_records_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageResult_AiOperationRecordRead__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Records",
        "tags": [
          "Code Review"
        ]
      }
    },
    "/api/code-review/records/{record_id}": {
      "get": {
        "operationId": "get_record_api_code_review_records__record_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AiOperationRecordRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Record",
        "tags": [
          "Code Review"
        ]
      }
    },
    "/api/code-review/review": {
      "post": {
        "operationId": "review_api_code_review_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_CodeReviewVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review",
        "tags": [
          "Code Review"
        ]
      }
    },
    "/api/knowledge/search": {
      "post": {
        "operationId": "search_api_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_KnowledgeSearchHit__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search",
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/upload": {
      "post": {
        "operationId": "upload_api_knowledge_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_KnowledgeDocumentVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload",
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/upload-file": {
      "post": {
        "operationId": "upload_file_api_knowledge_upload_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_knowledge_upload_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_KnowledgeDocumentVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "KNOWLEDGE_FILE_PARSE_FAILED"
                }
              }
            },
            "description": "文件为空、格式不支持、超限或解析失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload File",
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/{doc_id}": {
      "delete": {
        "operationId": "delete_api_knowledge__doc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete",
        "tags": [
          "Knowledge"
        ]
      }
    },
...
  }
}